REFACTOR `QueryBuilderTest` & `QueryBuilderWhereLikeTest` to 'tests/Feature/Builders' directory
